style: move styled-component to styles
